Move editor utility tests into tests folder
