+++ src/features/editor/tests/text-operations.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vite-plus/test";
-import { indentText, outdentText, toggleCaseText } from "./text-operations";
+import { indentText, outdentText, toggleCaseText } from "../utils/text-operations";
 
 describe("editor text operations", () => {
   it("inserts indentation at the cursor", () => {
